requirements.txt
# EmoTorch Requirements
# Python dependencies for emotion recognition models
# Created: August 2025

torch>=2.0.0
torchvision>=0.15.0
kornia>=0.7.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0